moved audio thread and functions to audio.c
cleared groupchat functions from list.c
